requirements.txt
Django==2.2.4
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
django-import-export==1.2.0
